Cleanup legacy transaction logics, old code unnecesary state
