+++ src/cow-react/modules/limitOrders/containers/LimitOrdersConfirmModal/index.tsx
@@ -6,7 +6,6 @@ import { CurrencyInfo } from '@cow/common/pure/CurrencyInputPanel/types'
 import { LimitOrdersConfirm } from '../../pure/LimitOrdersConfirm'
 import { TradeFlowContext } from '../../services/tradeFlow'
 import TransactionConfirmationModal, { OperationType } from 'components/TransactionConfirmationModal'
-import { L2Content as TxSubmittedModal } from 'components/TransactionConfirmationModal'
 import { limitOrdersConfirmState } from '../LimitOrdersConfirmModal/state'
 import { useWalletInfo } from '@cow/modules/wallet'
 import { GpModal } from '@cow/common/pure/Modal'
@@ -25,6 +24,7 @@ import { limitRateAtom } from '@cow/modules/limitOrders/state/limitRateAtom'
 import { useHandleOrderPlacement } from '@cow/modules/limitOrders/hooks/useHandleOrderPlacement'
 import { partiallyFillableOverrideAtom } from '@cow/modules/limitOrders/state/partiallyFillableOverride'
 import { useFeatureFlags } from '@cow/common/hooks/useFeatureFlags'
+import { L2Content as TxSubmittedModal } from 'components/TransactionConfirmationModal'
 
 export interface LimitOrdersConfirmModalProps {
   isOpen: boolean
