+++ src/custom/components/TransactionConfirmationModal/TransactionConfirmationModalMod.tsx
@@ -2,25 +2,19 @@ import { Trans } from '@lingui/macro'
 import { Currency } from '@uniswap/sdk-core'
 import Badge from 'components/Badge'
 import { getChainInfo } from 'constants/chainInfo'
-// import { SupportedL2ChainId } from 'constants/chains'
-// import useCurrencyLogoURIs from 'lib/hooks/useCurrencyLogoURIs'
-import { ReactNode /*, useCallback*/, useContext /*, useState*/, useMemo } from 'react'
-import { AlertCircle, AlertTriangle /*, ArrowUpCircle, CheckCircle */ } from 'react-feather'
+import { ReactNode, useContext, useMemo } from 'react'
+import { AlertCircle, AlertTriangle } from 'react-feather'
 import { Text } from 'rebass'
-import { useIsTransactionConfirmed, useTransaction } from 'state/transactions/hooks'
 import styled, { ThemeContext } from 'styled-components/macro'
 import { isL2ChainId } from 'utils/chains'
 
 import Circle from 'assets/images/blue-loader.svg'
 import { CloseIcon, CustomLightSpinner, ExternalLink } from 'theme'
-// import { ExplorerDataType, getExplorerLink } from 'utils/getExplorerLink'
-import { /* ButtonLight */ ButtonPrimary } from '../Button'
+import { ButtonPrimary } from '../Button'
 import { AutoColumn, ColumnCenter } from 'components/Column'
-// import Modal from 'components/Modal'
 import { RowBetween, RowFixed } from 'components/Row'
 import AnimatedConfirmation from 'components/TransactionConfirmationModal/AnimatedConfirmation'
 
-// MOD imports
 import { GpModal } from '@cow/common/pure/Modal'
 import {
   ConfirmationPendingContent,
@@ -32,6 +26,7 @@ import {
 } from '.'
 import { SupportedChainId } from '@cowprotocol/cow-sdk'
 import { useUpdateAtom, handleFollowPendingTxPopupAtom } from 'state/application/atoms'
+import { useIsTransactionConfirmed, useTransaction } from 'state/enhancedTransactions/hooks'
 import { getEtherscanLink as getExplorerLink } from 'utils'
 import { useWalletInfo } from '@cow/modules/wallet'
 
@@ -73,156 +68,24 @@ const StyledLogo = styled.img`
   margin-left: 6px;
 `
 
-/* export function ConfirmationPendingContent({
-  onDismiss,
-  pendingText,
-  inline,
-}: {
-  onDismiss: () => void
-  pendingText: ReactNode
-  inline?: boolean // not in modal
-}) {
-  return (
-    <Wrapper>
-      <AutoColumn gap="md">
-        {!inline && (
-          <RowBetween>
-            <div />
-            <CloseIcon onClick={onDismiss} />
-          </RowBetween>
-        )}
-        <ConfirmedIcon inline={inline}>
-          <CustomLightSpinner src={Circle} alt="loader" size={inline ? '40px' : '90px'} />
-        </ConfirmedIcon>
-        <AutoColumn gap="12px" justify={'center'}>
-          <Text fontWeight={500} fontSize={20} textAlign="center">
-            <Trans>Waiting For Confirmation</Trans>
-          </Text>
-          <Text fontWeight={400} fontSize={16} textAlign="center">
-            {pendingText}
-          </Text>
-          <Text fontWeight={500} fontSize={14} color="#565A69" textAlign="center" marginBottom="12px">
-            <Trans>Confirm this transaction in your wallet</Trans>
-          </Text>
-        </AutoColumn>
-      </AutoColumn>
-    </Wrapper>
-  )
-}
-export function TransactionSubmittedContent({
-  onDismiss,
-  chainId,
-  hash,
-  currencyToAdd,
-  inline,
-}: {
-  onDismiss: () => void
-  hash: string | undefined
-  chainId: number
-  currencyToAdd?: Currency | undefined
-  inline?: boolean // not in modal
-}) {
-  const theme = useContext(ThemeContext)
-
-  const { connector } = useWeb3React()
-
-  const token = currencyToAdd?.wrapped
-  const logoURL = useCurrencyLogoURIs(token)[0]
-
-  const [success, setSuccess] = useState<boolean | undefined>()
-
-  const addToken = useCallback(() => {
-    if (!token?.symbol || !connector.watchAsset) return
-    connector
-      .watchAsset({
-        address: token.address,
-        symbol: token.symbol,
-        decimals: token.decimals,
-        image: logoURL,
-      })
-      .then(() => setSuccess(true))
-      .catch(() => setSuccess(false))
-  }, [connector, logoURL, token])
-
-  return (
-    <Wrapper>
-      <Section inline={inline}>
-        {!inline && (
-          <RowBetween>
-            <div />
-            <CloseIcon onClick={onDismiss} />
-          </RowBetween>
-        )}
-        <ConfirmedIcon inline={inline}>
-          <ArrowUpCircle strokeWidth={0.5} size={inline ? '40px' : '90px'} color={theme.primary1} />
-        </ConfirmedIcon>
-        <AutoColumn gap="12px" justify={'center'}>
-          <Text fontWeight={500} fontSize={20} textAlign="center">
-            <Trans>Transaction Submitted</Trans>
-          </Text>
-          {chainId && hash && (
-            <ExternalLink href={getExplorerLink(chainId, hash, ExplorerDataType.TRANSACTION)}>
-              <Text fontWeight={500} fontSize={14} color={theme.primary1}>
-                <Trans>View on Explorer</Trans>
-              </Text>
-            </ExternalLink>
-          )}
-          {currencyToAdd && connector.watchAsset && (
-            <ButtonLight mt="12px" padding="6px 12px" width="fit-content" onClick={addToken}>
-              {!success ? (
-                <RowFixed>
-                  <Trans>Add {currencyToAdd.symbol}</Trans>
-                </RowFixed>
-              ) : (
-                <RowFixed>
-                  <Trans>Added {currencyToAdd.symbol} </Trans>
-                  <CheckCircle size={'16px'} stroke={theme.green1} style={{ marginLeft: '6px' }} />
-                </RowFixed>
-              )}
-            </ButtonLight>
-          )}
-          <ButtonPrimary onClick={onDismiss} style={{ margin: '20px 0 0 0' }}>
-            <Text fontWeight={500} fontSize={20}>
-              {inline ? <Trans>Return</Trans> : <Trans>Close</Trans>}
-            </Text>
-          </ButtonPrimary>
-        </AutoColumn>
-      </Section>
-    </Wrapper>
-  )
-} */
-
 export function ConfirmationModalContent({
   title,
-  titleSize, // mod
-  styles, // mod
-  className, // mod
+  titleSize,
+  styles,
+  className,
   bottomContent,
   onDismiss,
   topContent,
 }: ConfirmationModalContentProps) {
-  /* {
-  title: ReactNode
-  onDismiss: () => void
-  topContent: () => ReactNode
-  bottomContent?: () => ReactNode | undefined
-} */
   return (
     <Wrapper className={className}>
       <Section>
-        {/* <RowBetween> */}
         <GPModalHeader>
-          <Text
-            fontWeight={600} // MOD
-            fontSize={titleSize || 16} // MOD
-            style={styles} //MOD
-          >
+          <Text fontWeight={600} fontSize={titleSize || 16} style={styles}>
             {title}
           </Text>
-          {/* <CloseIcon onClick={onDismiss} /> */}
-          <CloseIconWrapper onClick={onDismiss} /> {/* MOD */}
+          <CloseIconWrapper onClick={onDismiss} />
         </GPModalHeader>
-        {/* </RowBetween> */}
         {topContent()}
       </Section>
       {bottomContent && <BottomSection gap="12px">{bottomContent()}</BottomSection>}
@@ -327,11 +190,7 @@ export function L2Content({
           </RowBetween>
         )}
         <ConfirmedIcon inline={inline}>{displayIcon}</ConfirmedIcon>
-        <AutoColumn
-          gap="12px"
-          justify={'center'}
-          style={{ margin: '0 0 16px' }} // MOD
-        >
+        <AutoColumn gap="12px" justify={'center'} style={{ margin: '0 0 16px' }}>
           <Text fontWeight={500} fontSize={20} textAlign="center">
             {!hash ? (
               <Trans>Confirm transaction in wallet</Trans>
@@ -381,12 +240,12 @@ export function L2Content({
 export interface ConfirmationModalProps {
   isOpen: boolean
   onDismiss: () => void
-  hash?: string | undefined // mod
-  content?: () => ReactNode // mod
+  hash?: string | undefined
+  content?: () => ReactNode
   attemptingTxn: boolean
   pendingText: ReactNode
   currencyToAdd?: Currency | undefined
-  operationType: OperationType // mod
+  operationType: OperationType
 }
 
 export default function TransactionConfirmationModal({
@@ -397,7 +256,7 @@ export default function TransactionConfirmationModal({
   pendingText,
   content,
   currencyToAdd,
-  operationType, // mod
+  operationType,
 }: ConfirmationModalProps) {
   const { chainId } = useWalletInfo()
   const setShowFollowPendingTxPopup = useUpdateAtom(handleFollowPendingTxPopupAtom)
@@ -414,7 +273,6 @@ export default function TransactionConfirmationModal({
 
   // confirmation screen
   return (
-    // <Modal isOpen={isOpen} onDismiss={onDismiss} maxHeight={90}>
     <GpModal isOpen={isOpen} onDismiss={_onDismiss} maxHeight={90} maxWidth={hash ? 850 : 470}>
       {isL2ChainId(chainId) && (hash || attemptingTxn) ? (
         <L2Content chainId={chainId} hash={hash} onDismiss={onDismiss} pendingText={pendingText} />
