+++ src/index.tsx
@@ -21,7 +21,7 @@ import './cow-react'
 // import ApplicationUpdater from 'state/application/updater'
 // import ListsUpdater from 'state/lists/updater'
 // import LogsUpdater from 'state/logs/updater'
-// import TransactionUpdater from 'state/transactions/updater'
+// import TransactionUpdater from 'state/enhancedTransactions/updater'
 // import UserUpdater from 'state/user/updater'
 // import GnosisSafeUpdater from 'state/gnosisSafe/updater'
 // import ThemeProvider, { FixedGlobalStyle, ThemedGlobalStyle } from 'theme'
