--- src/state/governance/hooks.ts
@@ -1,549 +1,0 @@
-import { useWalletInfo } from '@cow/modules/wallet'
-import { defaultAbiCoder, Interface } from '@ethersproject/abi'
-import { isAddress } from '@ethersproject/address'
-import { BigNumber } from '@ethersproject/bignumber'
-import { Contract } from '@ethersproject/contracts'
-import { TransactionResponse } from '@ethersproject/providers'
-import { toUtf8String, Utf8ErrorFuncs, Utf8ErrorReason } from '@ethersproject/strings'
-// eslint-disable-next-line no-restricted-imports
-import { t } from '@lingui/macro'
-import { abi as GOVERNANCE_ABI } from '@uniswap/governance/build/GovernorAlpha.json'
-import { abi as UNI_ABI } from '@uniswap/governance/build/Uni.json'
-import { CurrencyAmount, Token } from '@uniswap/sdk-core'
-import { useWeb3React } from '@web3-react/core'
-import GOVERNOR_BRAVO_ABI from 'abis/governor-bravo.json'
-import {
-  GOVERNANCE_ALPHA_V0_ADDRESSES,
-  GOVERNANCE_ALPHA_V1_ADDRESSES,
-  GOVERNANCE_BRAVO_ADDRESSES,
-} from 'constants/addresses'
-import { LATEST_GOVERNOR_INDEX } from 'constants/governance'
-import { POLYGON_PROPOSAL_TITLE } from 'constants/proposals/polygon_proposal_title'
-import { UNISWAP_GRANTS_PROPOSAL_DESCRIPTION } from 'constants/proposals/uniswap_grants_proposal_description'
-import { useContract } from 'hooks/useContract'
-import { useSingleCallResult, useSingleContractMultipleData } from 'lib/hooks/multicall'
-import { useCallback, useMemo } from 'react'
-import { calculateGasMargin } from 'utils/calculateGasMargin'
-import { SupportedChainId } from '../../constants/chains'
-import {
-  BRAVO_START_BLOCK,
-  MOONBEAN_START_BLOCK,
-  ONE_BIP_START_BLOCK,
-  POLYGON_START_BLOCK,
-  UNISWAP_GRANTS_START_BLOCK,
-} from '../../constants/proposals'
-import { UNI } from '../../constants/tokens'
-import { useLogs } from '../logs/hooks'
-import { useTransactionAdder } from '../transactions/hooks'
-import { TransactionType } from '../transactions/types'
-import { VoteOption } from './types'
-
-function useGovernanceV0Contract(): Contract | null {
-  return useContract(GOVERNANCE_ALPHA_V0_ADDRESSES, GOVERNANCE_ABI, false)
-}
-
-function useGovernanceV1Contract(): Contract | null {
-  return useContract(GOVERNANCE_ALPHA_V1_ADDRESSES, GOVERNANCE_ABI, false)
-}
-
-function useGovernanceBravoContract(): Contract | null {
-  return useContract(GOVERNANCE_BRAVO_ADDRESSES, GOVERNOR_BRAVO_ABI, true)
-}
-
-const useLatestGovernanceContract = useGovernanceBravoContract
-
-export function useUniContract() {
-  const { chainId } = useWalletInfo()
-  const uniAddress = useMemo(() => (chainId ? UNI[chainId]?.address : undefined), [chainId])
-  return useContract(uniAddress, UNI_ABI, true)
-}
-
-interface ProposalDetail {
-  target: string
-  functionSig: string
-  callData: string
-}
-
-export interface ProposalData {
-  id: string
-  title: string
-  description: string
-  proposer: string
-  status: ProposalState
-  forCount: CurrencyAmount<Token>
-  againstCount: CurrencyAmount<Token>
-  startBlock: number
-  endBlock: number
-  eta: BigNumber
-  details: ProposalDetail[]
-  governorIndex: number // index in the governance address array for which this proposal pertains
-}
-
-export interface CreateProposalData {
-  targets: string[]
-  values: string[]
-  signatures: string[]
-  calldatas: string[]
-  description: string
-}
-
-export enum ProposalState {
-  UNDETERMINED = -1,
-  PENDING,
-  ACTIVE,
-  CANCELED,
-  DEFEATED,
-  SUCCEEDED,
-  QUEUED,
-  EXPIRED,
-  EXECUTED,
-}
-
-const GovernanceInterface = new Interface(GOVERNANCE_ABI)
-
-// get count of all proposals made in the latest governor contract
-function useProposalCount(contract: Contract | null): number | undefined {
-  const { result } = useSingleCallResult(contract, 'proposalCount')
-
-  return result?.[0]?.toNumber()
-}
-
-interface FormattedProposalLog {
-  description: string
-  details: { target: string; functionSig: string; callData: string }[]
-}
-
-const FOUR_BYTES_DIR: { [sig: string]: string } = {
-  '0x5ef2c7f0': 'setSubnodeRecord(bytes32,bytes32,address,address,uint64)',
-  '0x10f13a8c': 'setText(bytes32,string,string)',
-  '0xb4720477': 'sendMessageToChild(address,bytes)',
-}
-
-/**
- * Need proposal events to get description data emitted from
- * new proposal event.
- */
-function useFormattedProposalCreatedLogs(
-  contract: Contract | null,
-  indices: number[][],
-  fromBlock?: number,
-  toBlock?: number
-): FormattedProposalLog[] | undefined {
-  // create filters for ProposalCreated events
-  const filter = useMemo(() => {
-    const filter = contract?.filters?.ProposalCreated()
-    if (!filter) return undefined
-    return {
-      ...filter,
-      fromBlock,
-      toBlock,
-    }
-  }, [contract, fromBlock, toBlock])
-
-  const useLogsResult = useLogs(filter)
-
-  return useMemo(() => {
-    return useLogsResult?.logs
-      ?.map((log) => {
-        const parsed = GovernanceInterface.parseLog(log).args
-        return parsed
-      })
-      ?.filter((parsed) => indices.flat().some((i) => i === parsed.id.toNumber()))
-      ?.map((parsed) => {
-        let description!: string
-
-        const startBlock = parseInt(parsed.startBlock?.toString())
-        try {
-          description = parsed.description
-        } catch (error: any) {
-          // replace invalid UTF-8 in the description with replacement characters
-          let onError = Utf8ErrorFuncs.replace
-
-          // Bravo proposal reverses the codepoints for U+2018 (‘) and U+2026 (…)
-          if (startBlock === BRAVO_START_BLOCK) {
-            const U2018 = [0xe2, 0x80, 0x98].toString()
-            const U2026 = [0xe2, 0x80, 0xa6].toString()
-            onError = (reason, offset, bytes, output) => {
-              if (reason === Utf8ErrorReason.UNEXPECTED_CONTINUE) {
-                const charCode = [bytes[offset], bytes[offset + 1], bytes[offset + 2]].reverse().toString()
-                if (charCode === U2018) {
-                  output.push(0x2018)
-                  return 2
-                } else if (charCode === U2026) {
-                  output.push(0x2026)
-                  return 2
-                }
-              }
-              return Utf8ErrorFuncs.replace(reason, offset, bytes, output)
-            }
-          }
-
-          description = JSON.parse(toUtf8String(error.error.value, onError)) || ''
-        }
-
-        // some proposals omit newlines
-        if (
-          startBlock === BRAVO_START_BLOCK ||
-          startBlock === ONE_BIP_START_BLOCK ||
-          startBlock === MOONBEAN_START_BLOCK
-        ) {
-          description = description.replace(/ {2}/g, '\n').replace(/\d\. /g, '\n$&')
-        }
-
-        return {
-          description,
-          details: parsed.targets.map((target: string, i: number) => {
-            const signature = parsed.signatures[i]
-            let calldata = parsed.calldatas[i]
-            let name: string
-            let types: string
-            if (signature === '') {
-              const fourbyte = calldata.slice(0, 10)
-              const sig = FOUR_BYTES_DIR[fourbyte] ?? 'UNKNOWN()'
-              if (!sig) throw new Error('Missing four byte sig')
-              ;[name, types] = sig.substring(0, sig.length - 1).split('(')
-              calldata = `0x${calldata.slice(10)}`
-            } else {
-              ;[name, types] = signature.substring(0, signature.length - 1).split('(')
-            }
-            const decoded = defaultAbiCoder.decode(types.split(','), calldata)
-            return {
-              target,
-              functionSig: name,
-              callData: decoded.join(', '),
-            }
-          }),
-        }
-      })
-  }, [indices, useLogsResult])
-}
-
-const V0_PROPOSAL_IDS = [[1], [2], [3], [4]]
-const V1_PROPOSAL_IDS = [[1], [2], [3]]
-
-function countToIndices(count: number | undefined, skip = 0) {
-  return typeof count === 'number' ? new Array(count - skip).fill(0).map((_, i) => [i + 1 + skip]) : []
-}
-
-// get data for all past and active proposals
-export function useAllProposalData(): { data: ProposalData[]; loading: boolean } {
-  const { chainId } = useWalletInfo()
-  const gov0 = useGovernanceV0Contract()
-  const gov1 = useGovernanceV1Contract()
-  const gov2 = useGovernanceBravoContract()
-
-  const proposalCount0 = useProposalCount(gov0)
-  const proposalCount1 = useProposalCount(gov1)
-  const proposalCount2 = useProposalCount(gov2)
-
-  const gov0ProposalIndexes = useMemo(() => {
-    return chainId === SupportedChainId.MAINNET ? V0_PROPOSAL_IDS : countToIndices(proposalCount0)
-  }, [chainId, proposalCount0])
-  const gov1ProposalIndexes = useMemo(() => {
-    return chainId === SupportedChainId.MAINNET ? V1_PROPOSAL_IDS : countToIndices(proposalCount1)
-  }, [chainId, proposalCount1])
-  const gov2ProposalIndexes = useMemo(() => {
-    return countToIndices(proposalCount2, 8)
-  }, [proposalCount2])
-
-  const proposalsV0 = useSingleContractMultipleData(gov0, 'proposals', gov0ProposalIndexes)
-  const proposalsV1 = useSingleContractMultipleData(gov1, 'proposals', gov1ProposalIndexes)
-  const proposalsV2 = useSingleContractMultipleData(gov2, 'proposals', gov2ProposalIndexes)
-
-  // get all proposal states
-  const proposalStatesV0 = useSingleContractMultipleData(gov0, 'state', gov0ProposalIndexes)
-  const proposalStatesV1 = useSingleContractMultipleData(gov1, 'state', gov1ProposalIndexes)
-  const proposalStatesV2 = useSingleContractMultipleData(gov2, 'state', gov2ProposalIndexes)
-
-  // get metadata from past events
-  const formattedLogsV0 = useFormattedProposalCreatedLogs(gov0, gov0ProposalIndexes, 11042287, 12563484)
-  const formattedLogsV1 = useFormattedProposalCreatedLogs(gov1, gov1ProposalIndexes, 12686656, 13059343)
-  const formattedLogsV2 = useFormattedProposalCreatedLogs(gov2, gov2ProposalIndexes, 13538153)
-
-  const uni = useMemo(() => (chainId ? UNI[chainId] : undefined), [chainId])
-
-  // early return until events are fetched
-  return useMemo(() => {
-    const proposalsCallData = [...proposalsV0, ...proposalsV1, ...proposalsV2]
-    const proposalStatesCallData = [...proposalStatesV0, ...proposalStatesV1, ...proposalStatesV2]
-    const formattedLogs = [...(formattedLogsV0 ?? []), ...(formattedLogsV1 ?? []), ...(formattedLogsV2 ?? [])]
-
-    if (
-      !uni ||
-      proposalsCallData.some((p) => p.loading) ||
-      proposalStatesCallData.some((p) => p.loading) ||
-      (gov0 && !formattedLogsV0) ||
-      (gov1 && !formattedLogsV1) ||
-      (gov2 && !formattedLogsV2)
-    ) {
-      return { data: [], loading: true }
-    }
-
-    return {
-      data: proposalsCallData.map((proposal, i) => {
-        const startBlock = parseInt(proposal?.result?.startBlock?.toString())
-
-        let description = formattedLogs[i]?.description ?? ''
-        if (startBlock === UNISWAP_GRANTS_START_BLOCK) {
-          description = UNISWAP_GRANTS_PROPOSAL_DESCRIPTION
-        }
-
-        let title = description?.split(/#+\s|\n/g)[1]
-        if (startBlock === POLYGON_START_BLOCK) {
-          title = POLYGON_PROPOSAL_TITLE
-        }
-
-        return {
-          id: proposal?.result?.id.toString(),
-          title: title ?? t`Untitled`,
-          description: description ?? t`No description.`,
-          proposer: proposal?.result?.proposer,
-          status: proposalStatesCallData[i]?.result?.[0] ?? ProposalState.UNDETERMINED,
-          forCount: CurrencyAmount.fromRawAmount(uni, proposal?.result?.forVotes),
-          againstCount: CurrencyAmount.fromRawAmount(uni, proposal?.result?.againstVotes),
-          startBlock,
-          endBlock: parseInt(proposal?.result?.endBlock?.toString()),
-          eta: proposal?.result?.eta,
-          details: formattedLogs[i]?.details,
-          governorIndex: i >= proposalsV0.length + proposalsV1.length ? 2 : i >= proposalsV0.length ? 1 : 0,
-        }
-      }),
-      loading: false,
-    }
-  }, [
-    formattedLogsV0,
-    formattedLogsV1,
-    formattedLogsV2,
-    gov0,
-    gov1,
-    gov2,
-    proposalStatesV0,
-    proposalStatesV1,
-    proposalStatesV2,
-    proposalsV0,
-    proposalsV1,
-    proposalsV2,
-    uni,
-  ])
-}
-
-export function useProposalData(governorIndex: number, id: string): ProposalData | undefined {
-  const { data } = useAllProposalData()
-  return data.filter((p) => p.governorIndex === governorIndex)?.find((p) => p.id === id)
-}
-
-export function useQuorum(governorIndex: number): CurrencyAmount<Token> | undefined {
-  const latestGovernanceContract = useLatestGovernanceContract()
-  const quorumVotes = useSingleCallResult(latestGovernanceContract, 'quorumVotes')?.result?.[0]
-  const { chainId } = useWalletInfo()
-  const uni = useMemo(() => (chainId ? UNI[chainId] : undefined), [chainId])
-
-  if (
-    !latestGovernanceContract ||
-    !quorumVotes ||
-    chainId !== SupportedChainId.MAINNET ||
-    !uni ||
-    governorIndex !== LATEST_GOVERNOR_INDEX
-  )
-    return undefined
-
-  return CurrencyAmount.fromRawAmount(uni, quorumVotes)
-}
-
-// get the users delegatee if it exists
-export function useUserDelegatee(): string {
-  const { account } = useWalletInfo()
-  const uniContract = useUniContract()
-  const { result } = useSingleCallResult(uniContract, 'delegates', [account ?? undefined])
-  return result?.[0] ?? undefined
-}
-
-// gets the users current votes
-export function useUserVotes(): { loading: boolean; votes: CurrencyAmount<Token> | undefined } {
-  const { account, chainId } = useWalletInfo()
-  const uniContract = useUniContract()
-
-  // check for available votes
-  const { result, loading } = useSingleCallResult(uniContract, 'getCurrentVotes', [account ?? undefined])
-  return useMemo(() => {
-    const uni = chainId ? UNI[chainId] : undefined
-    return { loading, votes: uni && result ? CurrencyAmount.fromRawAmount(uni, result?.[0]) : undefined }
-  }, [chainId, loading, result])
-}
-
-// fetch available votes as of block (usually proposal start block)
-export function useUserVotesAsOfBlock(block: number | undefined): CurrencyAmount<Token> | undefined {
-  const { account, chainId } = useWalletInfo()
-  const uniContract = useUniContract()
-
-  // check for available votes
-  const uni = useMemo(() => (chainId ? UNI[chainId] : undefined), [chainId])
-  const votes = useSingleCallResult(uniContract, 'getPriorVotes', [account ?? undefined, block ?? undefined])
-    ?.result?.[0]
-  return votes && uni ? CurrencyAmount.fromRawAmount(uni, votes) : undefined
-}
-
-export function useDelegateCallback(): (delegatee: string | undefined) => undefined | Promise<string> {
-  const { account, chainId } = useWalletInfo()
-  const { provider } = useWeb3React()
-  const addTransaction = useTransactionAdder()
-
-  const uniContract = useUniContract()
-
-  return useCallback(
-    (delegatee: string | undefined) => {
-      if (!provider || !chainId || !account || !delegatee || !isAddress(delegatee ?? '')) return undefined
-      const args = [delegatee]
-      if (!uniContract) throw new Error('No UNI Contract!')
-      return uniContract.estimateGas.delegate(...args, {}).then((estimatedGasLimit) => {
-        return uniContract
-          .delegate(...args, { value: null, gasLimit: calculateGasMargin(estimatedGasLimit) })
-          .then((response: TransactionResponse) => {
-            addTransaction(response, {
-              type: TransactionType.DELEGATE,
-              delegatee,
-            })
-            return response.hash
-          })
-      })
-    },
-    [account, addTransaction, chainId, provider, uniContract]
-  )
-}
-
-export function useVoteCallback(): (
-  proposalId: string | undefined,
-  voteOption: VoteOption
-) => undefined | Promise<string> {
-  const { account, chainId } = useWalletInfo()
-  const latestGovernanceContract = useLatestGovernanceContract()
-  const addTransaction = useTransactionAdder()
-
-  return useCallback(
-    (proposalId: string | undefined, voteOption: VoteOption) => {
-      if (!account || !latestGovernanceContract || !proposalId || !chainId) return
-      const args = [proposalId, voteOption === VoteOption.Against ? 0 : voteOption === VoteOption.For ? 1 : 2]
-      return latestGovernanceContract.estimateGas.castVote(...args, {}).then((estimatedGasLimit) => {
-        return latestGovernanceContract
-          .castVote(...args, { value: null, gasLimit: calculateGasMargin(estimatedGasLimit) })
-          .then((response: TransactionResponse) => {
-            addTransaction(response, {
-              type: TransactionType.VOTE,
-              decision: voteOption,
-              governorAddress: latestGovernanceContract.address,
-              proposalId: parseInt(proposalId),
-              reason: '',
-            })
-            return response.hash
-          })
-      })
-    },
-    [account, addTransaction, latestGovernanceContract, chainId]
-  )
-}
-
-export function useQueueCallback(): (proposalId: string | undefined) => undefined | Promise<string> {
-  const { account, chainId } = useWalletInfo()
-  const latestGovernanceContract = useLatestGovernanceContract()
-  const addTransaction = useTransactionAdder()
-
-  return useCallback(
-    (proposalId: string | undefined) => {
-      if (!account || !latestGovernanceContract || !proposalId || !chainId) return
-      const args = [proposalId]
-      return latestGovernanceContract.estimateGas.queue(...args, {}).then((estimatedGasLimit) => {
-        return latestGovernanceContract
-          .queue(...args, { value: null, gasLimit: calculateGasMargin(estimatedGasLimit) })
-          .then((response: TransactionResponse) => {
-            addTransaction(response, {
-              type: TransactionType.QUEUE,
-              governorAddress: latestGovernanceContract.address,
-              proposalId: parseInt(proposalId),
-            })
-            return response.hash
-          })
-      })
-    },
-    [account, addTransaction, latestGovernanceContract, chainId]
-  )
-}
-
-export function useExecuteCallback(): (proposalId: string | undefined) => undefined | Promise<string> {
-  const { account, chainId } = useWalletInfo()
-  const latestGovernanceContract = useLatestGovernanceContract()
-  const addTransaction = useTransactionAdder()
-
-  return useCallback(
-    (proposalId: string | undefined) => {
-      if (!account || !latestGovernanceContract || !proposalId || !chainId) return
-      const args = [proposalId]
-      return latestGovernanceContract.estimateGas.execute(...args, {}).then((estimatedGasLimit) => {
-        return latestGovernanceContract
-          .execute(...args, { value: null, gasLimit: calculateGasMargin(estimatedGasLimit) })
-          .then((response: TransactionResponse) => {
-            addTransaction(response, {
-              type: TransactionType.EXECUTE,
-              governorAddress: latestGovernanceContract.address,
-              proposalId: parseInt(proposalId),
-            })
-            return response.hash
-          })
-      })
-    },
-    [account, addTransaction, latestGovernanceContract, chainId]
-  )
-}
-
-export function useCreateProposalCallback(): (
-  createProposalData: CreateProposalData | undefined
-) => undefined | Promise<string> {
-  const { account, chainId } = useWalletInfo()
-  const latestGovernanceContract = useLatestGovernanceContract()
-  const addTransaction = useTransactionAdder()
-
-  return useCallback(
-    (createProposalData: CreateProposalData | undefined) => {
-      if (!account || !latestGovernanceContract || !createProposalData || !chainId) return undefined
-
-      const args = [
-        createProposalData.targets,
-        createProposalData.values,
-        createProposalData.signatures,
-        createProposalData.calldatas,
-        createProposalData.description,
-      ]
-
-      return latestGovernanceContract.estimateGas.propose(...args).then((estimatedGasLimit) => {
-        return latestGovernanceContract
-          .propose(...args, { gasLimit: calculateGasMargin(estimatedGasLimit) })
-          .then((response: TransactionResponse) => {
-            addTransaction(response, {
-              type: TransactionType.SUBMIT_PROPOSAL,
-            })
-            return response.hash
-          })
-      })
-    },
-    [account, addTransaction, latestGovernanceContract, chainId]
-  )
-}
-
-export function useLatestProposalId(address: string | undefined): string | undefined {
-  const latestGovernanceContract = useLatestGovernanceContract()
-  const res = useSingleCallResult(latestGovernanceContract, 'latestProposalIds', [address])
-  return res?.result?.[0]?.toString()
-}
-
-export function useProposalThreshold(): CurrencyAmount<Token> | undefined {
-  const { chainId } = useWalletInfo()
-
-  const latestGovernanceContract = useLatestGovernanceContract()
-  const res = useSingleCallResult(latestGovernanceContract, 'proposalThreshold')
-  const uni = useMemo(() => (chainId ? UNI[chainId] : undefined), [chainId])
-
-  if (res?.result?.[0] && uni) {
-    return CurrencyAmount.fromRawAmount(uni, res.result[0])
-  }
-
-  return undefined
-}
