--- src/state/governance/types.ts
@@ -1,5 +1,0 @@
-export enum VoteOption {
-  Against,
-  For,
-  Abstain,
-}
