--- src/state/transactions/hooks.tsx
@@ -1,83 +1,0 @@
-import { useWalletInfo } from '@cow/modules/wallet'
-import { TransactionResponse } from '@ethersproject/providers'
-import { useCallback, useMemo } from 'react'
-import { useAppDispatch } from '@src/custom/state/hooks'
-
-import { addTransaction } from './reducer'
-import { TransactionInfo } from './types'
-import { useAllTransactions } from '@src/custom/state/enhancedTransactions/hooks'
-import { EnhancedTransactionDetails } from '@src/custom/state/enhancedTransactions/reducer'
-
-// helper that can take a ethers library transaction response and add it to the list of transactions
-export function useTransactionAdder(): (response: TransactionResponse, info: TransactionInfo) => void {
-  const { chainId, account } = useWalletInfo()
-  const dispatch = useAppDispatch()
-
-  return useCallback(
-    (response: TransactionResponse, info: TransactionInfo) => {
-      if (!account) return
-      if (!chainId) return
-
-      const { hash } = response
-      if (!hash) {
-        throw Error('No transaction hash found.')
-      }
-      dispatch(addTransaction({ hash, from: account, info, chainId }))
-    },
-    [account, chainId, dispatch]
-  )
-}
-
-export function useTransaction(transactionHash?: string): EnhancedTransactionDetails | undefined {
-  const allTransactions = useAllTransactions()
-
-  if (!transactionHash) {
-    return undefined
-  }
-
-  return allTransactions[transactionHash]
-}
-
-export function useIsTransactionPending(transactionHash?: string): boolean {
-  const transactions = useAllTransactions()
-
-  if (!transactionHash || !transactions[transactionHash]) return false
-
-  return !transactions[transactionHash].receipt
-}
-
-export function useIsTransactionConfirmed(transactionHash?: string): boolean {
-  const transactions = useAllTransactions()
-
-  if (!transactionHash || !transactions[transactionHash]) return false
-
-  return Boolean(transactions[transactionHash].receipt)
-}
-
-/**
- * Returns whether a transaction happened in the last day (86400 seconds * 1000 milliseconds / second)
- * @param tx to check for recency
- */
-export function isTransactionRecent(tx: EnhancedTransactionDetails): boolean {
-  return new Date().getTime() - tx.addedTime < 86_400_000
-}
-
-// watch for submissions to claim
-// return null if not done loading, return undefined if not found
-export function useUserHasSubmittedClaim(account?: string): {
-  claimSubmitted: boolean
-  claimTxn: EnhancedTransactionDetails | undefined
-} {
-  const allTransactions = useAllTransactions()
-
-  // get the txn if it has been submitted
-  const claimTxn = useMemo(() => {
-    const txnIndex = Object.keys(allTransactions).find((hash) => {
-      const tx = allTransactions[hash]
-      return tx.claim && tx.claim.recipient === account
-    })
-    return txnIndex && allTransactions[txnIndex] ? allTransactions[txnIndex] : undefined
-  }, [account, allTransactions])
-
-  return { claimSubmitted: Boolean(claimTxn), claimTxn }
-}
